Adição de lógica de autorização no sistema utilizando spring security

+++ src/main/java/com/ufape/agiota/comunication/controllers/AgiotaController.java
@@ -5,6 +5,11 @@ import com.ufape.agiota.negocio.models.Agiota;
 import org.modelmapper.ModelMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.*;
 
 import com.ufape.agiota.comunication.dto.agiota.AgiotaRequest;
@@ -23,24 +28,31 @@ public class AgiotaController {
 
     @PostMapping("/agiota")
     public ResponseEntity<AgiotaResponse> saveAgiota(@Valid @RequestBody AgiotaRequest agiota){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Jwt principal = (Jwt) authentication.getPrincipal();
+        if (facade.checkDuplicatedUser(principal.getSubject())){throw new AccessDeniedException("User already exists");}
         AgiotaResponse response = new AgiotaResponse(facade.saveAgiota(agiota.convertToEntity()));
         return new ResponseEntity<>(response, HttpStatus.CREATED);
     }
 
     @PatchMapping("/agiota/{id}")
     public ResponseEntity<AgiotaResponse> updateAgiota(@PathVariable Long id, @Valid @RequestBody AgiotaRequest entity){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Jwt principal = (Jwt) authentication.getPrincipal();
         Agiota agiota = facade.findAgiota(id);
+        if(!principal.getSubject().equals(agiota.getIdKc())){throw new AccessDeniedException("User not allowed");}
         modelMapper.map(entity, agiota);
         AgiotaResponse response = new AgiotaResponse(facade.saveAgiota(agiota));
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
-    
+    @PreAuthorize("hasAnyRole('ADMINISTRADOR', 'AGIOTA', 'CUSTOMER')")
     @GetMapping("/agiota/{id}")
     public ResponseEntity<AgiotaResponse> findAgiota(@PathVariable Long id){
         AgiotaResponse response = new AgiotaResponse(facade.findAgiota(id));
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
+    @PreAuthorize("hasAnyRole('ADMINISTRADOR', 'AGIOTA', 'CUSTOMER')")
     @GetMapping("/agiotas")
     public ResponseEntity<Iterable<AgiotaResponse>> findAllAgiotas(){
         Iterable<AgiotaResponse> response = facade.findAllAgiotas().stream().map(AgiotaResponse::new).toList();
@@ -49,6 +61,9 @@ public class AgiotaController {
 
     @DeleteMapping("/agiota/{id}")
     public ResponseEntity<String> deleteAgiota(@PathVariable Long id){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Jwt principal = (Jwt) authentication.getPrincipal();
+        if(!principal.getSubject().equals(facade.findAgiota(id).getIdKc())){throw new AccessDeniedException("User not allowed");}
         facade.deleteAgiota(id);
         return new ResponseEntity<>("Agiota com ID " + id + " foi deletado", HttpStatus.NO_CONTENT);
     }
