+++ src/main/java/com/ufape/agiota/comunication/controllers/CustomerController.java
@@ -9,6 +9,11 @@ import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -18,17 +23,22 @@ public class CustomerController {
     final private Facade facade;
 
     @PostMapping
-    public ResponseEntity<CustomerResponse> saveCustomer(@Valid @RequestBody CustomerRequest customer){
-       CustomerResponse response = new CustomerResponse(facade.saveCustomer(customer.convertToEntity()));
-       return new ResponseEntity<>(response, HttpStatus.CREATED);
+    public ResponseEntity<CustomerResponse> saveCustomer(@Valid @RequestBody CustomerRequest customer) throws AccessDeniedException{
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Jwt principal = (Jwt) authentication.getPrincipal();
+        if (facade.checkDuplicatedUser(principal.getSubject())){throw new AccessDeniedException("User already exists");}
+        CustomerResponse response = new CustomerResponse(facade.saveCustomer(customer.convertToEntity()));
+        return new ResponseEntity<>(response, HttpStatus.CREATED);
     }
 
+    @PreAuthorize("hasAnyRole('ADMINISTRADOR', 'AGIOTA', 'CUSTOMER')")
     @GetMapping("/{id}")
     public ResponseEntity<CustomerResponse> findCustomer(@PathVariable Long id){
         CustomerResponse response = new CustomerResponse(facade.findCustomer(id));
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
+    @PreAuthorize("hasAnyRole('ADMINISTRADOR', 'AGIOTA', 'CUSTOMER')")
     @GetMapping
     public ResponseEntity<Iterable<CustomerResponse>> findAllCustomers(){
         Iterable<CustomerResponse> response = facade.findAllCustomers().stream().map(CustomerResponse::new).toList();
@@ -37,7 +47,10 @@ public class CustomerController {
 
     @PatchMapping("/{id}")
     public ResponseEntity<CustomerResponse> updateCustomer(@PathVariable Long id, @Valid @RequestBody CustomerRequest entity){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Jwt principal = (Jwt) authentication.getPrincipal();
         Customer customer = facade.findCustomer(id);
+        if(!principal.getSubject().equals(customer.getIdKc())){throw new AccessDeniedException("User not allowed");}
         modelMapper.map(entity, customer);
         CustomerResponse response = new CustomerResponse(facade.saveCustomer(customer));
         return new ResponseEntity<>(response, HttpStatus.OK);
@@ -45,6 +58,9 @@ public class CustomerController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCustomer(@PathVariable Long id){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Jwt principal = (Jwt) authentication.getPrincipal();
+        if(!principal.getSubject().equals(facade.findCustomer(id).getIdKc())){throw new AccessDeniedException("User not allowed");}
         facade.deleteCustomer(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
